Fixed a ai bug.
Calculating the mean heigth for a block was done by dividing by an int,
but it should be done by a float. Also change ai from doing calculation
in float instead of double.

+++ src/ai.cpp
@@ -9,13 +9,13 @@
 bool loadAi(Ai& ai, std::string dir, std::string str) {
 	std::ifstream file(dir+str);
 	if (file.is_open()) {
-		std::vector<double> values;
+		std::vector<float> values;
 		while (file.good()) {
 			std::string line;
 			std::getline(file, line);
 			std::stringstream strStream(line);
 			std::string word;
-			double value;
+			float value;
 			strStream >> word >> value; // E.g. word = "rowRoughness". calue = 10.1 .
 			values.push_back(value);
 		}
