+++ src/computer.h
@@ -41,7 +41,7 @@ private:
 		int rotations_;
 	};
 
-	double calculateValue(const RawTetrisBoard& board, const Block&) const;
+	float calculateValue(const RawTetrisBoard& board, const Block&) const;
 
 	// Calculate and return the best input to achieve the current state.
 	Input calculateInput(State state) const;
